requirements.txt
Flask
face_recognition
numpy
Pillow
